consts & converters with tests
